backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
redis==5.0.1
requests==2.32.3
celery==5.3.4
prometheus-client==0.19.0
python-barcode==0.15.1
qrcode[pil]==7.4.2
Pillow==10.1.0
# Optional: background scheduler (notifications cron)
schedule==1.2.0
